[mod] moved primary.tsx to button.tsx
